refactor: Retire use of `download_dir` as `TensorizerArgs` param
Replaces `download_dir` with `tensorizer_uri` as a `TensorizerArgs`
param. This is due to discussions on `download_dir` being a confusing
and ultimately not helpful parameter for the location of model tensors.
Instead, `download_dir` is back to the definition coinciding with
the convention set by HuggingFace; as a location to download weights
for caching. A new parameter takes its place: `tensorizer_uri`. This
specifically deals with locating model tensors for `tensorizer`.